backend/parser.py
import requests
from bs4 import BeautifulSoup
from typing import Dict, Any
import re

class Form144Parser:
    def __init__(self):
        self.headers = {
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36'
        }

    async def fetch_document(self, url: str) -> str:
        """Fetch the Form 144 document from EDGAR."""
        response = requests.get(url, headers=self.headers)
        response.raise_for_status()
        return response.text

    def parse_form144(self, content: str) -> Dict[str, Any]:
        """Parse Form 144 content and extract structured data."""
        soup = BeautifulSoup(content, 'html.parser')
        
        # Initialize structured data
        data = {
            'issuer': {},
            'filer': {},
            'transaction': {},
            'securities': {}
        }

        # Extract issuer information
        issuer_section = soup.find('div', string=re.compile('Issuer', re.I))
        if issuer_section:
            data['issuer'] = self._extract_issuer_info(issuer_section)

        # Extract filer information
        filer_section = soup.find('div', string=re.compile('Filer', re.I))
        if filer_section:
            data['filer'] = self._extract_filer_info(filer_section)

        # Extract transaction details
        transaction_section = soup.find('div', string=re.compile('Transaction', re.I))
        if transaction_section:
            data['transaction'] = self._extract_transaction_info(transaction_section)

        # Extract securities information
        securities_section = soup.find('div', string=re.compile('Securities', re.I))
        if securities_section:
            data['securities'] = self._extract_securities_info(securities_section)

        return data

    def _extract_issuer_info(self, section) -> Dict[str, str]:
        """Extract issuer information from the section."""
        info = {}
        # TODO: Implement issuer information extraction
        return info

    def _extract_filer_info(self, section) -> Dict[str, str]:
        """Extract filer information from the section."""
        info = {}
        # TODO: Implement filer information extraction
        return info

    def _extract_transaction_info(self, section) -> Dict[str, str]:
        """Extract transaction information from the section."""
        info = {}
        # TODO: Implement transaction information extraction
        return info

    def _extract_securities_info(self, section) -> Dict[str, str]:
        """Extract securities information from the section."""
        info = {}
        # TODO: Implement securities information extraction
        return info

    def generate_summary(self, data: Dict[str, Any]) -> str:
        """Generate a human-readable summary of the Form 144 filing."""
        summary = []
        
        if data['issuer']:
            summary.append(f"Issuer: {data['issuer'].get('name', 'N/A')}")
        
        if data['filer']:
            summary.append(f"Filer: {data['filer'].get('name', 'N/A')}")
        
        if data['transaction']:
            summary.append(f"Transaction Type: {data['transaction'].get('type', 'N/A')}")
            summary.append(f"Transaction Date: {data['transaction'].get('date', 'N/A')}")
        
        if data['securities']:
            summary.append(f"Securities Type: {data['securities'].get('type', 'N/A')}")
            summary.append(f"Number of Shares: {data['securities'].get('shares', 'N/A')}")
        
        return "\n".join(summary) 