backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
transformers==4.35.2
torch==2.1.1
faiss-cpu==1.7.4
langchain==0.0.350
python-multipart==0.0.6
aiohttp==3.9.1
pytest==7.4.3
httpx==0.25.2 